# rank-stars-api-docs.yaml
openapi: 3.0.4
info: 
  title: RankStars API
  description: |-
    This is the documentation of the RankStars API. You can access the website at [https://website.com](https://website.com).
    
    This documentation aims to provide all the necessary information to use and understand the API.
    
  contact:
    email: vitorlemes.com@hotmail.com
  version: 1.0.0
servers:
  - url: https://website.com/api/
...
paths:
...
  /reviews:
    get:
      tags: 
        - review
      summary: Get all reviews
      description: Get all the reviews matching the specified filters.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            example: "Some search"
        
        - name: sortBy
          in: query
          required: false
          schema: 
            type: string
            example: "createdAt"
            enum: [createdAt, updatedAt, rating]
        
        - name: order
          in: query
          required: false
          schema: 
            type: string
            example: "-1"
            
        - name: limit
          in: query
          required: false
          schema: 
            type: number
            example: "10"
            
        - name: skip
          in: query
          required: false
          schema: 
            type: number
            example: "20"
        
        - name: author
          in: query
          required: false
          description: Expects the author ID
          schema: 
            type: string
            example: "678ec07b6e268a81e2f3a687"
        
        - name: minRating
          in: query
          required: false
          schema: 
            type: number
            example: "60"
            
        - name: maxRating
          in: query
          required: false
          schema:
            type: number
            example: 100
      
        - name: minDate
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2024-01-01"
            
        - name: maxDate
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-31"
        - name: category
          in: query
          required: false
          description: Can receive multiple values separated by comma
          schema:
            type: string
            example: "tvshows,movies,videogames" 
            
      responses:
        "200":
          description: An array of reviews
        "400":
          description: Invalid value to query parameter.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: 
        - review
      summary: Create a new review
      description: The user needs to be authenticated to perform this action.
      requestBody:
        description: Review structure
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - rating
                - type
              properties:
                title:
                  type: string
                  example: New Review
                rating:
                  type: number
                  example: 100
                review:
                  type: string
                type:
                  type: string
                  example: tvshows
                image:
                  type: array
                  format: binary
                  items:
                    type: string
      responses:
        "200":
          description: Sucessful operation
        "401":
          description: User is not authenticated.
        "422":
          description: Validation error.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /lists/{listId}/reviews:
    get:
      tags: 
        - list
      summary: Get reviews from list
      description: Get the reviews from an specific list based on the id.
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            example: "678ec07b6e268a81e2f3a687"
        
        - name: populate
          in: query
          required: false
          description: Populate the reviews.
          schema: 
            type: boolean
      responses:
        "200":
          description: An array of reviews
        "404":
          description: List not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        data: 
          type: object